python_backend/requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
websockets>=12.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0
psycopg[binary]>=3.2.0
asyncpg>=0.29.0
alembic>=1.14.0

# Configuration
python-dotenv>=1.0.0